Basic admin system. This should automatically prevent most admin commands from being called by players that shouldn't be calling them, and even supports multiple levels of admins.
There are some outlier commands that don't follow the assumed convention of "admin" ability + godLevel, for example /setGodMode does not require the "admin" ability, but neither does it have a godLevel. So without manually checking in that instance, any player could  grant themselves god mode and have access to low level admin commands.

In order for admins to gain access to most of their commands, they must call the /setGodMode command. This command gives them the "admin" ability, and this ability works in tandem with their godLevel (which is determined by their access level).

Adds a minor implementation for abilities to enable the admin system to work. This implementation should come in handy in GU3 when we start on combat, since it will automatically prevent players from calling combat commands that they do not have the ability to call.